backend/requirements.txt
# FinSage Backend Requirements
# Run from: cd backend && pip install -r requirements.txt

# ── Core LangGraph / LangChain ──────────────────────────────────────────────
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# ── LLM Providers ────────────────────────────────────────────────────────────
langchain-google-genai>=2.0.0
langchain-openai>=0.2.0

# ── Backend / API ────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# ── Database ─────────────────────────────────────────────────────────────────
psycopg2-binary>=2.9.9

# ── Utilities ────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
rich>=13.0.0
pydantic>=2.0.0
typing-extensions>=4.9.0
